requirements.txt
pika==1.3.2
pika-stubs==0.1.3
requests==2.31.0
boto3==1.34.55
python-decouple==3.8
